Modified interaction with usable items. Begin add stuff for talkable items/scenery.
Added fading of critter/map text.
Modified algorithm for rectangle determination of items on hex.

Npc spawn reduced to 2 hexes radius, like be before.

ASCompiler modification of showed error/warning lines.

Minor fixes, refactoring.

git-svn-id: https://cvet.by:8443/svn/FOnlineSource@85 c2c6a35e-ad2b-4e1f-bb64-97b989be360b

Former-commit-id: 608d553adf1e5e95de130647ca432f2afa308ed8

+++ Source/Client.h
@@ -101,9 +101,9 @@ public:
 	void Net_SendLogIn(const char* name, const char* pass);
 	void Net_SendCreatePlayer(CritterCl* newcr);
 	void Net_SendSaveLoad(bool save, const char* fname, ByteVec* pic_data);
-	void Net_SendUseSkill(BYTE skill, CritterCl* cr);
-	void Net_SendUseSkill(BYTE skill, ItemHex* item);
-	void Net_SendUseSkill(BYTE skill, Item* item);
+	void Net_SendUseSkill(WORD skill, CritterCl* cr);
+	void Net_SendUseSkill(WORD skill, ItemHex* item);
+	void Net_SendUseSkill(WORD skill, Item* item);
 	void Net_SendUseItem(BYTE ap, DWORD item_id, WORD item_pid, BYTE rate, BYTE target_type, DWORD target_id, WORD target_pid, DWORD param);
 	void Net_SendPickItem(WORD targ_x, WORD targ_y, WORD pid);
 	void Net_SendPickCritter(DWORD crid, BYTE pick_type);
